test: clean up coverage follow-up tests
