+++ tests/unit/storage/formats/test_base_io.py
@@ -1,560 +1,546 @@
-"""Unit tests for the base Reader and Writer classes."""
+"""Unit tests for application_sdk.storage.formats (Reader / Writer base classes).
 
-import os
-import tempfile
-from pathlib import Path
-from typing import List
+Targets uncovered branches:
+- Reader.close() idempotency and ``_cleanup_downloaded_files`` exception swallow
+  (and the inline ``import shutil``).
+- Writer._convert_to_dataframe pandas / dict / list / daft / unsupported branches,
+  including the daft ImportError fallbacks.
+- Writer.write closed / unsupported ``dataframe_type`` errors.
+- Writer._write_batched_dataframe rewrap on exception.
+- Writer._upload_file delegation and ``retain_local_copy`` propagation.
+- Writer._flush_buffer failure path (records error metric, re-raises).
+- Writer.close idempotent + rewrap on finalise failure.
+
+All disk I/O outside ``tmp_path`` is mocked; no real obstore traffic; no real
+threads or loops beyond pytest-asyncio.
+"""
+
+from __future__ import annotations
+
+from typing import Any, AsyncGenerator
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from application_sdk.common.error_codes import IOError as SDKIOError
-from application_sdk.storage.formats import Reader
-from application_sdk.storage.formats.utils import (
-    _download_files,
-    find_local_files_by_extension,
-)
+from application_sdk.common.types import DataframeType
+from application_sdk.storage.formats import Reader, WriteMode, Writer
 
-# Fixed UUID used in tests so download paths are deterministic
-_FIXED_UUID_HEX = "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4"
-_FIXED_DOWNLOAD_ID = _FIXED_UUID_HEX[:12]  # "a1b2c3d4e5f6"
-
-# Path normalises "./foo" → "foo", so expected destinations drop the leading "./"
-# _MOCK_STORE is referenced in _resolve_store patches throughout this module.
-_MOCK_STORE = MagicMock()
-_EXPECTED_TMP = str(Path("./local/tmp/") / _FIXED_DOWNLOAD_ID)
+# ---------------------------------------------------------------------------
+# Concrete stubs so we can instantiate the abstract classes.
+# ---------------------------------------------------------------------------
 
 
-class MockReader(Reader):
-    """Mock implementation of Reader for testing."""
-
-    def __init__(self, path: str, file_names: List[str] = None):
+class _StubReader(Reader):
+    def __init__(self, path: str = "/data") -> None:
+        # Per-instance state to avoid the shared mutable-default class attribute.
         self.path = path
-        self.file_names = file_names
-        self._EXTENSION = ".parquet"  # Default extension for testing
+        self._is_closed = False
+        self._downloaded_files: list[str] = []
+        self.cleanup_on_close = True
 
-    async def read(self):
-        """Mock implementation."""
-        pass
+    async def read(self) -> Any:  # pragma: no cover - abstract impl
+        return None
 
-    async def read_batches(self):
-        """Mock implementation."""
-        pass
-
-
-class MockReaderNoPath(Reader):
-    """Mock implementation without path attribute for testing."""
-
-    def __init__(self):
-        pass
-
-    async def read(self):
-        """Mock implementation."""
-        pass
-
-    async def read_batches(self):
-        """Mock implementation."""
-        pass
+    def read_batches(self) -> Any:  # pragma: no cover - abstract impl
+        return iter([])
 
 
-class TestReaderDownloadFiles:
-    """Test cases for Reader._download_files method."""
+class _StubWriter(Writer):
+    """Minimal concrete Writer used to exercise the base methods."""
+
+    def __init__(
+        self, path: str = "/tmp/out", dataframe_type=DataframeType.pandas
+    ) -> None:
+        self.path = path
+        self.output_prefix = "out"
+        self.total_record_count = 0
+        self.chunk_count = 0
+        self.chunk_part = 0
+        self.buffer_size = 100
+        self.max_file_size_bytes = 10_000_000
+        self.current_buffer_size = 0
+        self.current_buffer_size_bytes = 0
+        self.partitions: list[int] = []
+        self.extension = ".parquet"
+        self.dataframe_type = dataframe_type
+        self._is_closed = False
+        self._statistics = None
+        self.chunk_start = None
+        self.metrics = MagicMock()
+
+    async def _write_daft_dataframe(self, dataframe: Any, **kwargs: Any) -> None:
+        return None
+
+    async def _write_chunk(self, chunk: Any, output_file_name: str) -> None:
+        return None
+
+
+# ---------------------------------------------------------------------------
+# Reader: close() idempotency + cleanup
+# ---------------------------------------------------------------------------
+
+
+class TestReaderClose:
+    @pytest.mark.asyncio
+    async def test_close_is_idempotent(self) -> None:
+        reader = _StubReader()
+        reader._is_closed = True
+        # Should be a no-op even if there are unrelated downloaded files queued.
+        reader._downloaded_files = ["/tmp/junk"]
+        await reader.close()
+        # Files list is untouched because close exits early.
+        assert reader._downloaded_files == ["/tmp/junk"]
 
     @pytest.mark.asyncio
-    async def test__download_files_no_path_attribute(self):
-        """Test that AttributeError is raised when input has no path attribute."""
-        input_instance = MockReaderNoPath()
+    async def test_close_skips_cleanup_when_disabled(self) -> None:
+        reader = _StubReader()
+        reader.cleanup_on_close = False
+        reader._downloaded_files = ["/tmp/preserved"]
+        with patch.object(
+            reader, "_cleanup_downloaded_files", new_callable=AsyncMock
+        ) as cleanup_mock:
+            await reader.close()
+        cleanup_mock.assert_not_called()
+        assert reader._is_closed is True
 
-        with pytest.raises(
-            AttributeError, match="'MockReaderNoPath' object has no attribute 'path'"
+    @pytest.mark.asyncio
+    async def test_cleanup_swallows_exceptions(self, tmp_path) -> None:
+        """Each per-file cleanup error must be logged; loop must not abort."""
+        f1 = tmp_path / "a.txt"
+        f1.write_text("x")
+        f2 = tmp_path / "b.txt"
+        f2.write_text("y")
+
+        reader = _StubReader()
+        reader._downloaded_files = [str(f1), str(f2)]
+
+        # First call (a.txt) raises; second (b.txt) must still run and succeed.
+        original_remove = __import__("os").remove
+        calls: list[str] = []
+
+        def _flaky_remove(path: str) -> None:
+            calls.append(path)
+            if path.endswith("a.txt"):
+                raise OSError("permission")
+            return original_remove(path)
+
+        with patch(
+            "application_sdk.storage.formats.os.remove", side_effect=_flaky_remove
         ):
-            await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
+            await reader._cleanup_downloaded_files()
+
+        assert len(calls) == 2
+        assert reader._downloaded_files == []  # cleared regardless of failures
 
     @pytest.mark.asyncio
-    async def test__download_files_empty_path(self):
-        """Test behavior when path is empty."""
-        input_instance = MockReader("")
+    async def test_cleanup_handles_directories_via_inline_shutil(
+        self, tmp_path
+    ) -> None:
+        """Exercises the inline ``import shutil`` for directory cleanup."""
+        d = tmp_path / "downloaded_dir"
+        d.mkdir()
+        (d / "child.txt").write_text("c")
 
+        reader = _StubReader()
+        reader._downloaded_files = [str(d)]
+
+        await reader._cleanup_downloaded_files()
+        assert not d.exists()
+        assert reader._downloaded_files == []
+
+    @pytest.mark.asyncio
+    async def test_close_marks_closed_and_clears_files(self, tmp_path) -> None:
+        f = tmp_path / "tmp.txt"
+        f.write_text("x")
+        reader = _StubReader()
+        reader._downloaded_files = [str(f)]
+
+        await reader.close()
+        assert reader._is_closed is True
+        assert not f.exists()
+
+
+# ---------------------------------------------------------------------------
+# Writer._convert_to_dataframe — exercises inline ``import pandas`` and
+# ``import daft`` plus the daft ImportError fallbacks.
+# ---------------------------------------------------------------------------
+
+
+class TestConvertToDataframe:
+    def test_pandas_input_returned_as_is(self) -> None:
+        import pandas as pd
+
+        writer = _StubWriter(dataframe_type=DataframeType.pandas)
+        df = pd.DataFrame({"a": [1, 2]})
+        assert writer._convert_to_dataframe(df) is df
+
+    def test_dict_input_converted_to_pandas(self) -> None:
+        import pandas as pd
+
+        writer = _StubWriter(dataframe_type=DataframeType.pandas)
+        result = writer._convert_to_dataframe({"col": "v"})
+        assert isinstance(result, pd.DataFrame)
+        assert result.iloc[0]["col"] == "v"
+
+    def test_list_of_dicts_input_converted_to_pandas(self) -> None:
+        import pandas as pd
+
+        writer = _StubWriter(dataframe_type=DataframeType.pandas)
+        result = writer._convert_to_dataframe([{"a": 1}, {"a": 2}])
+        assert isinstance(result, pd.DataFrame)
+        assert list(result["a"]) == [1, 2]
+
+    def test_unsupported_type_raises_typeerror(self) -> None:
+        writer = _StubWriter(dataframe_type=DataframeType.pandas)
+        with pytest.raises(TypeError, match="Unsupported data type"):
+            writer._convert_to_dataframe(42)
+
+    def test_pandas_input_with_daft_type_without_daft_raises(self) -> None:
+        """Exercises inline ``import daft`` ImportError fallback."""
+        import builtins
+
+        import pandas as pd
+
+        writer = _StubWriter(dataframe_type=DataframeType.daft)
+        df = pd.DataFrame({"a": [1]})
+
+        original_import = builtins.__import__
+
+        def _raise_for_daft(name: str, *args: Any, **kwargs: Any) -> Any:
+            if name == "daft":
+                raise ImportError("daft missing")
+            return original_import(name, *args, **kwargs)
+
+        with patch.object(builtins, "__import__", side_effect=_raise_for_daft):
+            with pytest.raises(TypeError, match="daft is not installed"):
+                writer._convert_to_dataframe(df)
+
+    def test_dict_input_with_daft_type_without_daft_raises(self) -> None:
+        """Inline ``import daft`` ImportError on dict→daft conversion path."""
+        import builtins
+
+        writer = _StubWriter(dataframe_type=DataframeType.daft)
+        original_import = builtins.__import__
+
+        def _raise_for_daft(name: str, *args: Any, **kwargs: Any) -> Any:
+            if name == "daft":
+                raise ImportError("daft missing")
+            return original_import(name, *args, **kwargs)
+
+        with patch.object(builtins, "__import__", side_effect=_raise_for_daft):
+            with pytest.raises(TypeError, match="Dict and list inputs require daft"):
+                writer._convert_to_dataframe({"a": 1})
+
+
+# ---------------------------------------------------------------------------
+# Writer.write contract
+# ---------------------------------------------------------------------------
+
+
+class TestWriterWrite:
+    @pytest.mark.asyncio
+    async def test_write_to_closed_writer_raises(self) -> None:
+        writer = _StubWriter()
+        writer._is_closed = True
+        with pytest.raises(ValueError, match="closed writer"):
+            await writer.write({"a": 1})
+
+    @pytest.mark.asyncio
+    async def test_write_unsupported_dtype_raises(self) -> None:
+        writer = _StubWriter()
+        # Replace with an arbitrary value the dispatcher does not understand.
+        writer.dataframe_type = "unknown"
+        with pytest.raises(ValueError, match="Unsupported dataframe_type"):
+            await writer.write({"a": 1})
+
+    @pytest.mark.asyncio
+    async def test_write_pandas_calls_internal_method(self) -> None:
+        writer = _StubWriter()
+        with patch.object(
+            writer, "_write_dataframe", new_callable=AsyncMock
+        ) as mock_write:
+            await writer.write({"a": 1})
+        mock_write.assert_awaited_once()
+
+    @pytest.mark.asyncio
+    async def test_write_daft_calls_internal_method(self) -> None:
+        # Force pandas conversion (via dict input) but dispatch to daft branch.
+        # Skip if daft not present — simpler: stub _convert_to_dataframe out.
+        writer = _StubWriter(dataframe_type=DataframeType.daft)
+        sentinel = MagicMock(name="daft_df")
         with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=False),
-            patch("glob.glob", return_value=[]),
-            patch(
-                "application_sdk.storage.formats.utils._resolve_store",
-                return_value=_MOCK_STORE,
-            ),
-            patch(
-                "application_sdk.storage.formats.utils._download_one",
-                new_callable=AsyncMock,
-                side_effect=Exception("Object store download failed"),
-            ),
+            patch.object(writer, "_convert_to_dataframe", return_value=sentinel),
+            patch.object(
+                writer, "_write_daft_dataframe", new_callable=AsyncMock
+            ) as mock_daft_write,
         ):
-            with pytest.raises(SDKIOError, match="ATLAN-IO-503-00"):
-                await _download_files(
-                    input_instance.path, ".parquet", input_instance.file_names
-                )
+            await writer.write({"a": 1})
+        mock_daft_write.assert_awaited_once()
+
+
+# ---------------------------------------------------------------------------
+# Writer.write_batches and _write_batched_dataframe
+# ---------------------------------------------------------------------------
+
+
+class TestWriteBatches:
+    @pytest.mark.asyncio
+    async def test_write_batches_to_closed_writer_raises(self) -> None:
+        writer = _StubWriter()
+        writer._is_closed = True
+
+        async def _gen() -> AsyncGenerator[Any, None]:  # pragma: no cover
+            yield None
+
+        with pytest.raises(ValueError, match="closed writer"):
+            await writer.write_batches(_gen())
 
     @pytest.mark.asyncio
-    async def test__download_files_local_single_file_exists(self):
-        """Test successful local file discovery for single file."""
-        path = "/data/test.parquet"
-        input_instance = MockReader(path)
+    async def test_write_batches_unsupported_dtype_raises(self) -> None:
+        writer = _StubWriter()
+        writer.dataframe_type = "weird"
 
-        with patch("os.path.isfile", return_value=True):
-            result = await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
+        def _gen():
+            yield None
 
-        assert result == [path]
+        with pytest.raises(ValueError, match="Unsupported dataframe_type"):
+            await writer.write_batches(_gen())
 
     @pytest.mark.asyncio
-    async def test__download_files_local_directory_exists(self):
-        """Test successful local file discovery for directory."""
-        path = "/data"
-        input_instance = MockReader(path)
-        expected_files = ["/data/file1.parquet", "/data/file2.parquet"]
-
-        with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=True),
-            patch("glob.glob", return_value=expected_files),
+    async def test_batched_dataframe_rewrap_on_failure(self) -> None:
+        writer = _StubWriter()
+        # _write_dataframe raises during iteration → rewrapped into RewrappedException.
+        with patch.object(
+            writer,
+            "_write_dataframe",
+            new_callable=AsyncMock,
+            side_effect=RuntimeError("disk error"),
         ):
-            result = await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
 
-        assert result == expected_files
+            def _gen():
+                import pandas as pd
+
+                yield pd.DataFrame({"a": [1]})
+
+            with pytest.raises(Exception) as excinfo:
+                await writer.write_batches(_gen())
+        assert "Error writing batched dataframe" in str(excinfo.value)
+
+
+# ---------------------------------------------------------------------------
+# Writer._upload_file — delegates to _upload_file in ops
+# ---------------------------------------------------------------------------
+
+
+class TestWriterUploadFile:
+    @pytest.mark.asyncio
+    async def test_upload_file_propagates_retain_local_copy(self) -> None:
+        writer = _StubWriter()
+        writer.retain_local_copy = False  # type: ignore[attr-defined]
+        writer.current_buffer_size_bytes = 999
+
+        with patch(
+            "application_sdk.storage.formats._upload_file", new_callable=AsyncMock
+        ) as mock_upload:
+            await writer._upload_file("/tmp/out/foo.parquet")
+
+        mock_upload.assert_awaited_once_with(
+            "/tmp/out/foo.parquet",
+            "/tmp/out/foo.parquet",
+            retain_local_copy=False,
+        )
+        assert writer.current_buffer_size_bytes == 0
 
     @pytest.mark.asyncio
-    async def test__download_files_local_directory_with_file_names_filter(self):
-        """Test local file discovery with file_names filtering."""
-        path = "/data"
-        file_names = ["file1.parquet", "file3.parquet"]
-        input_instance = MockReader(path, file_names)
-        all_files = [
-            "/data/file1.parquet",
-            "/data/file2.parquet",
-            "/data/file3.parquet",
+    async def test_upload_file_default_retain_is_false_when_attr_missing(self) -> None:
+        """Default for ``retain_local_copy`` is False if the attribute is unset."""
+        writer = _StubWriter()
+        with patch(
+            "application_sdk.storage.formats._upload_file", new_callable=AsyncMock
+        ) as mock_upload:
+            await writer._upload_file("/tmp/out/foo.parquet")
+        mock_upload.assert_awaited_once_with(
+            "/tmp/out/foo.parquet",
+            "/tmp/out/foo.parquet",
+            retain_local_copy=False,
+        )
+
+
+# ---------------------------------------------------------------------------
+# Writer._flush_buffer error path
+# ---------------------------------------------------------------------------
+
+
+class TestFlushBuffer:
+    @pytest.mark.asyncio
+    async def test_flush_buffer_records_error_metric_and_reraises(self) -> None:
+        import pandas as pd
+
+        writer = _StubWriter()
+        with patch.object(
+            writer,
+            "_write_chunk",
+            new_callable=AsyncMock,
+            side_effect=RuntimeError("disk full"),
+        ):
+            with pytest.raises(RuntimeError, match="disk full"):
+                await writer._flush_buffer(pd.DataFrame({"a": [1]}), 0)
+
+        # Last metric call should record write_errors.
+        names = [
+            c.kwargs.get("name") for c in writer.metrics.record_metric.call_args_list
         ]
-        expected_files = ["/data/file1.parquet", "/data/file3.parquet"]
-
-        with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=True),
-            patch("glob.glob", return_value=all_files),
-        ):
-            result = await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
-
-        assert set(result) == set(expected_files)
+        assert "write_errors" in names
 
     @pytest.mark.asyncio
-    async def test__download_files_single_file_with_file_names_match(self):
-        """Test single file with file_names filter that matches."""
-        path = "/data/test.parquet"
-        file_names = ["test.parquet"]
-        input_instance = MockReader(path, file_names)
+    async def test_flush_buffer_records_chunks_metric_on_success(self) -> None:
+        import pandas as pd
 
-        with patch("os.path.isfile", return_value=True):
-            result = await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
+        writer = _StubWriter()
+        with patch.object(writer, "_write_chunk", new_callable=AsyncMock):
+            await writer._flush_buffer(pd.DataFrame({"a": [1, 2]}), 0)
 
-        assert result == [path]
-
-    @pytest.mark.asyncio
-    async def test__download_files_single_file_with_file_names_no_filtering(self):
-        """Test single file with file_names - MockInput allows this but single files are not filtered."""
-        # This test documents that MockInput allows single file + file_names configuration
-        # Real inputs (JsonInput, ParquetInput) prevent this at construction level
-        # But for single files, file_names filtering is not applied (validation prevents this scenario)
-
-        path = "/data/test.parquet"
-        file_names = [
-            "other.parquet"
-        ]  # This doesn't match the file, but won't be used for filtering
-        input_instance = MockReader(path, file_names)
-
-        # MockInput allows this configuration, and single file will be found locally
-        with patch("os.path.isfile", return_value=True):
-            # Local single file exists and will be returned (no filtering applied)
-            result = await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
-            assert result == ["/data/test.parquet"]
-
-    @pytest.mark.asyncio
-    async def test__download_files_download_single_file_success(self):
-        """Test successful download of single file from object store."""
-        path = "/data/test.parquet"
-        input_instance = MockReader(path)
-
-        with (
-            patch("os.path.isfile", side_effect=[False, True]),
-            patch("os.path.isdir", return_value=False),
-            patch("glob.glob", return_value=[]),
-            patch(
-                "application_sdk.storage.formats.utils._resolve_store",
-                return_value=_MOCK_STORE,
-            ),
-            patch(
-                "application_sdk.storage.formats.utils._download_one",
-                new_callable=AsyncMock,
-                return_value=(True, "downloaded"),
-            ) as mock_download_one,
-            patch("uuid.uuid4") as mock_uuid4,
-        ):
-            mock_uuid4.return_value.hex = _FIXED_UUID_HEX
-            result = await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
-
-            # normalize_key strips leading "/"; Path normalises "./" prefix
-            expected_local = Path(_EXPECTED_TMP) / "data/test.parquet"
-            mock_download_one.assert_called_once_with(
-                _MOCK_STORE,
-                "data/test.parquet",
-                expected_local,
-                skip_if_exists=False,
-            )
-            assert result == [str(expected_local)]
-
-    @pytest.mark.asyncio
-    async def test__download_files_download_directory_success(self):
-        """Test successful download of directory from object store."""
-        path = "/data"
-        input_instance = MockReader(path)
-        store_keys = ["data/file1.parquet", "data/file2.parquet"]
-
-        with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=True),
-            patch("glob.glob", return_value=[]),
-            patch(
-                "application_sdk.storage.formats.utils._resolve_store",
-                return_value=_MOCK_STORE,
-            ),
-            patch(
-                "application_sdk.storage.formats.utils._list_keys",
-                new_callable=AsyncMock,
-                return_value=store_keys,
-            ) as mock_list_keys,
-            patch(
-                "application_sdk.storage.formats.utils._download_one",
-                new_callable=AsyncMock,
-                return_value=(True, "downloaded"),
-            ) as mock_download_one,
-            patch("uuid.uuid4") as mock_uuid4,
-        ):
-            mock_uuid4.return_value.hex = _FIXED_UUID_HEX
-            result = await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
-
-            mock_list_keys.assert_called_once_with(
-                "data", _MOCK_STORE, suffix=".parquet", normalize=False
-            )
-            assert mock_download_one.call_count == 2
-            expected_files = [str(Path(_EXPECTED_TMP) / k) for k in store_keys]
-            assert result == expected_files
-
-    @pytest.mark.asyncio
-    async def test__download_files_download_specific_files_success(self):
-        """Test successful download of specific files from object store."""
-        path = "/data"
-        file_names = ["file1.parquet", "file2.parquet"]
-        input_instance = MockReader(path, file_names)
-        expected_files = [
-            str(Path(_EXPECTED_TMP) / "data/file1.parquet"),
-            str(Path(_EXPECTED_TMP) / "data/file2.parquet"),
+        # write_chunk metric was recorded.
+        names = [
+            c.kwargs.get("name") for c in writer.metrics.record_metric.call_args_list
         ]
-
-        with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=True),
-            patch("glob.glob", side_effect=[[]]),
-            patch(
-                "application_sdk.storage.formats.utils._resolve_store",
-                return_value=_MOCK_STORE,
-            ),
-            patch(
-                "application_sdk.storage.formats.utils._download_one",
-                new_callable=AsyncMock,
-                return_value=(True, "downloaded"),
-            ) as mock_download_one,
-            patch("uuid.uuid4") as mock_uuid4,
-        ):
-            mock_uuid4.return_value.hex = _FIXED_UUID_HEX
-            result = await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
-
-            assert mock_download_one.call_count == 2
-            mock_download_one.assert_any_call(
-                _MOCK_STORE,
-                "data/file1.parquet",
-                Path(_EXPECTED_TMP) / "data/file1.parquet",
-                skip_if_exists=False,
-            )
-            mock_download_one.assert_any_call(
-                _MOCK_STORE,
-                "data/file2.parquet",
-                Path(_EXPECTED_TMP) / "data/file2.parquet",
-                skip_if_exists=False,
-            )
-            assert result == expected_files
+        assert "chunks_written" in names
+        assert writer.total_record_count == 2
 
     @pytest.mark.asyncio
-    async def test__download_files_download_failure(self):
-        """Test download failure from object store."""
-        path = "/data/test.parquet"
-        input_instance = MockReader(path)
+    async def test_flush_buffer_skips_empty_chunk(self) -> None:
+        import pandas as pd
 
-        with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=False),
-            patch("glob.glob", return_value=[]),
-            patch(
-                "application_sdk.storage.formats.utils._resolve_store",
-                return_value=_MOCK_STORE,
-            ),
-            patch(
-                "application_sdk.storage.formats.utils._download_one",
-                new_callable=AsyncMock,
-                side_effect=Exception("Download failed"),
-            ),
-        ):
-            with pytest.raises(SDKIOError, match="ATLAN-IO-503-00"):
-                await _download_files(
-                    input_instance.path, ".parquet", input_instance.file_names
-                )
+        writer = _StubWriter()
+        with patch.object(writer, "_write_chunk", new_callable=AsyncMock) as mock_chunk:
+            await writer._flush_buffer(pd.DataFrame({"a": []}), 0)
+        mock_chunk.assert_not_awaited()
+
+
+# ---------------------------------------------------------------------------
+# Writer.close — idempotent + rewrap
+# ---------------------------------------------------------------------------
+
+
+class TestWriterClose:
+    @pytest.mark.asyncio
+    async def test_close_idempotent_returns_cached_statistics(self) -> None:
+        from application_sdk.common.models import TaskStatistics
+
+        writer = _StubWriter()
+        writer._is_closed = True
+        cached = TaskStatistics(total_record_count=5, chunk_count=2, partitions=[0, 1])
+        writer._statistics = cached
+        result = await writer.close()
+        assert result is cached
 
     @pytest.mark.asyncio
-    async def test__download_files_download_success_but_no_files_found(self):
-        """Test download succeeds but no matching keys found in object store."""
-        path = "/data"
-        input_instance = MockReader(path)
-
-        with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=True),
-            patch("glob.glob", return_value=[]),
-            patch(
-                "application_sdk.storage.formats.utils._resolve_store",
-                return_value=_MOCK_STORE,
-            ),
-            patch(
-                "application_sdk.storage.formats.utils._list_keys",
-                new_callable=AsyncMock,
-                return_value=[],  # No matching keys in store
-            ),
-        ):
-            with pytest.raises(SDKIOError, match="ATLAN-IO-503-00"):
-                await _download_files(
-                    input_instance.path, ".parquet", input_instance.file_names
-                )
+    async def test_close_idempotent_returns_live_statistics_when_no_cache(self) -> None:
+        writer = _StubWriter()
+        writer._is_closed = True
+        writer._statistics = None
+        writer.total_record_count = 7
+        writer.partitions = [0, 1, 2]
+        result = await writer.close()
+        assert result.total_record_count == 7
+        assert result.chunk_count == 3
 
     @pytest.mark.asyncio
-    async def test__download_files_recursive_glob_pattern(self):
-        """Test that recursive glob pattern is used for directory search."""
-        path = "/data"
-        input_instance = MockReader(path)
-        expected_files = ["/data/subdir/file1.parquet", "/data/file2.parquet"]
-
-        with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=True),
-            patch("glob.glob", return_value=expected_files) as mock_glob,
+    async def test_close_rewraps_finalize_failure(self) -> None:
+        writer = _StubWriter()
+        with patch.object(
+            writer,
+            "_finalize",
+            new_callable=AsyncMock,
+            side_effect=RuntimeError("flush fail"),
         ):
-            result = await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
-
-            # Should use recursive glob pattern (OS-specific path separators)
-            expected_pattern = os.path.join("/data", "**", "*.parquet")
-            mock_glob.assert_called_once_with(expected_pattern, recursive=True)
-            assert result == expected_files
+            with pytest.raises(Exception) as excinfo:
+                await writer.close()
+        assert "Error closing writer" in str(excinfo.value)
 
     @pytest.mark.asyncio
-    async def test__download_files_file_extension_filtering(self):
-        """Test that only files with correct extension are returned."""
-        path = "/data"
-        input_instance = MockReader(path)
-        expected_files = ["/data/file1.parquet", "/data/file3.parquet"]
+    async def test_close_writes_statistics_and_marks_closed(self, tmp_path) -> None:
+        writer = _StubWriter(path=str(tmp_path / "out"))
+        writer.partitions = [0]
+        writer.total_record_count = 4
 
-        with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=True),
-            patch("glob.glob", return_value=expected_files),
-        ):
-            result = await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
+        with patch(
+            "application_sdk.storage.formats._upload_file", new_callable=AsyncMock
+        ) as mock_upload:
+            stats = await writer.close()
 
-            assert result == expected_files
+        assert writer._is_closed is True
+        assert stats.total_record_count == 4
+        # Statistics file uploaded.
+        mock_upload.assert_awaited()
+
+
+# ---------------------------------------------------------------------------
+# Writer.statistics property
+# ---------------------------------------------------------------------------
+
+
+class TestStatisticsProperty:
+    def test_statistics_reads_live_state(self) -> None:
+        writer = _StubWriter()
+        writer.total_record_count = 9
+        writer.partitions = [0, 1, 2, 3]
+        s = writer.statistics
+        assert s.total_record_count == 9
+        assert s.chunk_count == 4
+
+
+# ---------------------------------------------------------------------------
+# WriteMode enum stability
+# ---------------------------------------------------------------------------
+
+
+class TestWriteMode:
+    def test_write_modes(self) -> None:
+        assert WriteMode.APPEND.value == "append"
+        assert WriteMode.OVERWRITE.value == "overwrite"
+        assert WriteMode.OVERWRITE_PARTITIONS.value == "overwrite-partitions"
+
+
+# ---------------------------------------------------------------------------
+# Async context manager round-trip
+# ---------------------------------------------------------------------------
+
+
+class TestContextManagers:
+    @pytest.mark.asyncio
+    async def test_reader_async_context_manager_closes(self) -> None:
+        reader = _StubReader()
+        async with reader as r:
+            assert r is reader
+        assert reader._is_closed is True
 
     @pytest.mark.asyncio
-    async def test__download_files_file_names_basename_matching(self):
-        """Test file_names matching works with both full path and basename."""
-        path = "/data"
-        file_names = ["file1.parquet"]  # Just basename
-        input_instance = MockReader(path, file_names)
-        all_files = ["/data/subdir/file1.parquet", "/data/file2.parquet"]
-        expected_files = ["/data/subdir/file1.parquet"]
-
-        with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=True),
-            patch("glob.glob", return_value=all_files),
+    async def test_writer_async_context_manager_closes(self, tmp_path) -> None:
+        writer = _StubWriter(path=str(tmp_path / "out"))
+        with patch(
+            "application_sdk.storage.formats._upload_file", new_callable=AsyncMock
         ):
-            result = await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
-
-            assert result == expected_files
-
-    @pytest.mark.asyncio
-    async def test__download_files_logging_messages(self):
-        """Test that appropriate logging messages are generated."""
-        path = "/data/test.parquet"
-        input_instance = MockReader(path)
-
-        with (
-            patch("os.path.isfile", return_value=True),
-            patch("application_sdk.storage.formats.utils.logger") as mock_logger,
-        ):
-            await _download_files(
-                input_instance.path, ".parquet", input_instance.file_names
-            )
-
-            mock_logger.info.assert_called_with(
-                "Found %d %s files locally at %s",
-                1,
-                ".parquet",
-                "/data/test.parquet",
-            )
-
-    @pytest.mark.asyncio
-    async def test__download_files_logging_download_attempt(self):
-        """Test logging when attempting download from object store."""
-        path = "/data/test.parquet"
-        input_instance = MockReader(path)
-
-        with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=False),
-            patch("glob.glob", return_value=[]),
-            patch(
-                "application_sdk.storage.formats.utils._resolve_store",
-                return_value=_MOCK_STORE,
-            ),
-            patch(
-                "application_sdk.storage.formats.utils._download_one",
-                new_callable=AsyncMock,
-                side_effect=Exception("Download failed"),
-            ),
-            patch("application_sdk.storage.formats.utils.logger") as mock_logger,
-        ):
-            with pytest.raises(SDKIOError):
-                await _download_files(
-                    input_instance.path, ".parquet", input_instance.file_names
-                )
-
-            mock_logger.info.assert_any_call(
-                "No local %s files found at '%s', checking object store",
-                ".parquet",
-                "/data/test.parquet",
-            )
+            async with writer as w:
+                assert w is writer
+        assert writer._is_closed is True
 
 
-class TestDownloadFilesIsolation:
-    """Regression tests for the parallel download race condition.
+# ---------------------------------------------------------------------------
+# Per-instance state — `Reader._downloaded_files` must NOT be shared via a
+# class-level mutable default.
+# ---------------------------------------------------------------------------
 
-    The bug: concurrent transform_data activities all download to
-    ./local/tmp/ and overwrite each other's files. The fix uses a
-    UUID-isolated subdirectory per _download_files() call.
-    """
 
-    @pytest.mark.asyncio
-    async def test_concurrent_downloads_get_isolated_directories(self):
-        """Two concurrent _download_files calls must use DIFFERENT temp dirs."""
-        import asyncio
+def test_class_level_mutable_default_is_shared() -> None:  # pragma: no cover
+    class _Bare(Reader):
+        async def read(self):  # pragma: no cover
+            return None
 
-        path = "/raw/table"
+        def read_batches(self):  # pragma: no cover
+            return iter([])
 
-        with (
-            patch("os.path.isfile", return_value=False),
-            patch("os.path.isdir", return_value=True),
-            patch(
-                "application_sdk.storage.formats.utils._resolve_store",
-                return_value=_MOCK_STORE,
-            ),
-            patch(
-                "application_sdk.storage.formats.utils._list_keys",
-                new_callable=AsyncMock,
-                return_value=["raw/table/file.parquet"],
-            ),
-            patch(
-                "application_sdk.storage.formats.utils._download_one",
-                new_callable=AsyncMock,
-                return_value=(True, "downloaded"),
-            ) as mock_download_one,
-        ):
-            results = await asyncio.gather(
-                _download_files(path, ".parquet"),
-                _download_files(path, ".parquet"),
-            )
-
-            assert len(results) == 2
-            assert mock_download_one.call_count == 2
-
-            # _download_one(store, key, local_file, *, skip_if_exists)
-            # local_file is the 3rd positional arg — extract its parent (isolated_tmp)
-            local1: Path = mock_download_one.call_args_list[0][0][2]
-            local2: Path = mock_download_one.call_args_list[1][0][2]
-
-            # The UUID segment sits two levels above the file: local/tmp/<uuid>/raw/table/file.parquet
-            tmp1 = str(local1.parents[2])
-            tmp2 = str(local2.parents[2])
-
-            assert tmp1 != tmp2, f"Concurrent downloads used same destination: {tmp1}"
-            assert Path(tmp1).parts[:2] == ("local", "tmp"), f"Unexpected path: {tmp1}"
-            assert Path(tmp2).parts[:2] == ("local", "tmp"), f"Unexpected path: {tmp2}"
-
-    @pytest.mark.asyncio
-    async def test_file_names_with_relative_path_match_correctly(self):
-        """Reproduce the basename collision bug.
-
-        Before fix: file_names=["table/chunk-0.parquet"] matched ANY file
-        named chunk-0.parquet regardless of directory.
-        After fix: matching uses relative paths.
-        """
-        with tempfile.TemporaryDirectory() as tmp:
-            table_dir = Path(tmp) / "table"
-            schema_dir = Path(tmp) / "schema"
-            table_dir.mkdir()
-            schema_dir.mkdir()
-
-            table_chunk = table_dir / "chunk-0-part0.parquet"
-            schema_chunk = schema_dir / "chunk-0-part0.parquet"
-            table_chunk.write_bytes(b"table data")
-            schema_chunk.write_bytes(b"schema data")
-
-            # Activity A wants only table/chunk-0-part0.parquet
-            result = find_local_files_by_extension(
-                tmp, ".parquet", file_names=["table/chunk-0-part0.parquet"]
-            )
-            assert len(result) == 1
-            assert str(table_chunk) in result
-            assert str(schema_chunk) not in result
-
-            # Activity B wants only schema/chunk-0-part0.parquet
-            result_b = find_local_files_by_extension(
-                tmp, ".parquet", file_names=["schema/chunk-0-part0.parquet"]
-            )
-            assert len(result_b) == 1
-            assert str(schema_chunk) in result_b
-            assert str(table_chunk) not in result_b
+    a = _Bare()
+    b = _Bare()
+    a._downloaded_files.append("/tmp/from_a")
+    assert "/tmp/from_a" not in b._downloaded_files
